REST - save new order

+++ src/main/java/com/epam/pizzaservice/service/CustomerService.java
@@ -3,5 +3,5 @@ package com.epam.pizzaservice.service;
 import com.epam.pizzaservice.domain.Customer;
 
 public interface CustomerService {
-	Customer getCustomerById(Integer id);
+	Customer getCustomerById(Long id);
 }
